requirements.txt
nextcord
aiohttp # Could you tell me why we need three HTTP libraries? Seems unnecessary.
httpx
requests
Pillow
git+https://github.com/KaityXD/mafic # We should probably make this a PR.
git+https://github.com/nextcord/nextcord@master
colorama
psutil